Added two new tables to SINAN Added examples of using new tables to notebook Moved get_cid10_table to pysus.online_data package

+++ pysus/online_data/__init__.py
@@ -8,6 +8,7 @@ from pathlib import Path
 from ftplib import FTP, error_perm
 from dbfread import DBF
 from pysus.utilities.readdbc import read_dbc
+import pandas as pd
 
 CACHEPATH = os.getenv("PYSUS_CACHEPATH", os.path.join(str(Path.home()), "pysus"))
 
@@ -25,7 +26,7 @@ def cache_contents():
     return [os.path.join(CACHEPATH, f) for f in cached_data]
 
 
-def _fetch_file(fname, path, ftype):
+def _fetch_file(fname: str, path: str, ftype: str) -> pd.DataFrame:
     """
     Fetch a single file.
     :return:
@@ -43,5 +44,24 @@ def _fetch_file(fname, path, ftype):
     elif ftype == "DBF":
         dbf = DBF(fname, encoding="iso-8859-1")
         df = pd.DataFrame(list(dbf))
-    os.unlink(fname)
+
+    if os.path.exists(fname):
+        os.unlink(fname)
+    return df
+
+
+def get_CID10_table(cache=True):
+    """
+    Fetch the CID10 table
+    :param cache:
+    :return:
+    """
+    fname = "CID10.DBF"
+    cachefile = os.path.join(CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet")
+    if os.path.exists(cachefile):
+        df = pd.read_parquet(cachefile)
+        return df
+    df = _fetch_file(fname, "/dissemin/publicos/SIM/CID10/TABELAS", "DBF")
+    if cache:
+        df.to_parquet(cachefile)
     return df
